requirements.txt
streamlit
binance-connector
pandas
